Changed the size

+++ interfaces/IES1Benchmark.h
@@ -32,7 +32,7 @@ namespace Exchange {
     struct EXTERNAL IES1Benchmark : virtual public Core::IUnknown {
         enum { ID = ID_ES1BENCHMARK };
 
-        // using IUInt32Iterator = RPC::IIteratorType<uint32_t, ID_ES1BENCHMARK_UINT32ITERATOR>;
+        // using IUInt8Iterator = RPC::IIteratorType<uint8_t, ID_ES1BENCHMARK_UINT8ITERATOR>;
 
         ~IES1Benchmark() override = default;
 
@@ -40,7 +40,7 @@ namespace Exchange {
         virtual uint32_t EchoString(const string& value /* @in */, string& echo /* @out */) = 0;
 
         // @text echoarray
-        virtual uint32_t EchoArray(const std::vector<uint32_t>& values /* @in */, std::vector<uint32_t>& echo /* @out */) = 0;
+        virtual uint32_t EchoArray(const std::vector<uint8_t>& values /* @in @restrict:0..256K */, std::vector<uint8_t>& echo /* @out @restrict:0..256K */) = 0;
 
         // @text echoint32
         virtual uint32_t EchoUint32(const uint32_t value /* @in */, uint32_t& echo /* @out */) = 0;
